refactor: 클래스명 변경
dto로 사용됨에 따라 클래스명을 Dto를 붙여 변경

Related to: #7
