+++ guidely/src/main/java/dotori/guidely/domain/oauth/dto/AuthTokenDto.java
@@ -1,4 +1,4 @@
-package dotori.guidely.domain.oauth;
+package dotori.guidely.domain.oauth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,14 +9,14 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class AuthTokens {
+public class AuthTokenDto {
     private String accessToken;
     private String refreshToken;
     private String grantType;
     private Long expiresIn;
 
-    public static AuthTokens of(String accessToken, String refreshToken, String grantType, Long expiresIn) {
-        return new AuthTokens(accessToken, refreshToken, grantType, expiresIn);
+    public static AuthTokenDto of(String accessToken, String refreshToken, String grantType, Long expiresIn) {
+        return new AuthTokenDto(accessToken, refreshToken, grantType, expiresIn);
     }
 }
